Order Model and Cloudinary file upload

+++ controllers/cart.controller.js
@@ -6,6 +6,10 @@ const { successRes, catchRes, errorRes } = require("../utils/response.function")
 const addToCart = async (req, res) => {
     try {
         const cartBody = req.body
+        if (req.user.role !== 'buyer') {
+            return errorRes(res, "only buyer add the product into the cart")
+        }
+
         const findExistCartProduct = await cartService.findCart(cartBody)
         if (findExistCartProduct) {
             return errorRes(res, "user add this already into the cart")
@@ -23,6 +27,10 @@ const addToCart = async (req, res) => {
 const getCurrentUserCart = async (req, res) => {
     try {
         const { user_id } = req.user.id
+
+        if (req.user.role !== 'buyer') {
+            return errorRes(res, "only buyer can get our cart")
+        }
 
         const cartFound = await cartService.getAllCartItems(user_id)
 
@@ -42,6 +50,11 @@ const updateQuantity = async (req, res) => {
     try {
         const product_id = req.params.product_id
         const quantity = req.body
+
+        if (req.user.role !== 'buyer') {
+            return errorRes(res, "only buyer can update the quantity of the product")
+        }
+
         const data = await cartService.updateQuantity(quantity, { product_id: product_id })
         return successRes(res, "update quantity of cart", data, 200)
     }
@@ -55,6 +68,10 @@ const removeProductFromCart = async (req, res) => {
     try {
         const product_id = req.params.product_id
 
+        if (req.user.role !== 'buyer') {
+            return errorRes(res, "only buyer product can remove product from the cart")
+        }
+
         const data = await cartService.deleteProductFromCart({ product_id: product_id })
         return successRes(res, "remove product from cart", data, 200)
     }
@@ -67,6 +84,10 @@ const removeProductFromCart = async (req, res) => {
 const clearCartOfUser = async (req, res) => {
     try {
         const { user_id } = req.body;
+
+        if (req.user.role !== 'buyer') {
+            return errorRes(res, "only buyer clear our cart")
+        }
 
         if (!user_id) {
             return errorRes(res, "user_id is required", 400);
